production_requirements.txt
# DataGuardian Pro Production Requirements - Conflict Free
# =======================================================

# Core Web Framework
streamlit>=1.28.0

# AI/ML Dependencies  
anthropic>=0.7.0
openai>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Database & Caching
psycopg2-binary>=2.9.7
redis>=4.6.0

# HTTP & Web Scraping
requests>=2.31.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
tldextract>=3.4.0

# Document Processing (NO TEXTRACT - conflict resolved)
pypdf2>=3.0.1
reportlab>=4.0.0
pdfkit>=1.0.0

# Security & Authentication - COMPLETE SET
bcrypt>=4.0.1
pyjwt>=2.8.0
cryptography>=41.0.0
authlib>=1.2.1

# Payment Processing
stripe>=6.0.0

# Visualization & Plotting
plotly>=5.17.0

# System Monitoring & Performance
psutil>=5.9.0
memory-profiler>=0.60.0
cachetools>=5.3.0

# Utilities & Configuration
dnspython>=2.4.0
pyyaml>=6.0.1

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0

# OCR and Image Processing (REPLACES textract functionality)
pytesseract>=0.3.10
opencv-python-headless>=4.8.0

# Enterprise Dependencies
python-jose>=3.3.0
python-multipart>=0.0.6
joblib>=1.3.0

# Text Processing & Documents
markdown2>=2.4.0

# Document Generation & Export
weasyprint>=60.0
svglib>=1.5.1

# Web Framework Extensions
flask>=2.3.0

# Security & Identity
python-whois>=0.8.0
pycryptodome>=3.19.0

# Enterprise Authentication
python3-saml>=1.15.0

# Database Support (Multi-vendor)
mysql-connector-python>=8.2.0
psycopg2-pool>=1.1

# Enterprise AI/ML Processing
onnx>=1.15.0
onnxruntime>=1.16.0

# Advanced ML (Optional - comment out if causing issues)
# torch>=2.1.0
# tensorflow>=2.14.0

# Enterprise Monitoring & Profiling
py-spy>=0.3.14

# Additional document processing (Alternative to textract)
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.1

# File type detection
python-magic>=0.4.27
